// cognite/pygen/_typescript/instance_api/types/responses.ts
export interface InstanceResultItem {
  /**
   * The type of the instance (node or edge).
   */
  instanceType: InstanceType;

  /**
   * The space of the instance.
   */
  space: string;

  /**
   * The external ID of the instance.
   */
  externalId: string;

  /**
   * The version of the instance after the operation.
   */
  version: number;

  /**
   * Whether the instance was modified by the operation.
   */
  wasModified: boolean;

  /**
   * The time the instance was created (milliseconds since epoch).
   */
  createdTime: number;

  /**
   * The time the instance was last updated (milliseconds since epoch).
   */
  lastUpdatedTime: number;
}
// ...
export interface UpsertResult {
  /**
   * List of all instance result items from the operation.
   */
  items: InstanceResultItem[];

  /**
   * List of instance IDs that were deleted.
   */
  deleted: InstanceId[];
}
// ...
export interface DeleteResponse {
  /**
   * List of instance IDs that were deleted.
   */
  items: InstanceId[];
}